+++ src/app.d.ts
@@ -1,11 +1,11 @@
 /// <reference types="@sveltejs/kit" />
 
 declare module '*?geometrize' {
-	const placeholder: import('$lib/core/types.js').GeometrizePlaceholder;
+	const placeholder: import('#lib/core/types.js').GeometrizePlaceholder;
 	export default placeholder;
 }
 
 declare module '*&geometrize' {
-	const placeholder: import('$lib/core/types.js').GeometrizePlaceholder;
+	const placeholder: import('#lib/core/types.js').GeometrizePlaceholder;
 	export default placeholder;
 }